+++ lib/src/caches/ephemeral_fifo_cache.dart
@@ -74,6 +74,18 @@ class EphemeralFIFOCache<K, V> extends ThreadSafeCache<K, V> {
     });
   }
 
+  /// Removes the entry with the given key from the cache.
+  ///
+  /// - If the key does not exist, this call is a no-op.
+  ///
+  /// **This method is thread-safe.**
+  @override
+  Future<void> remove(K key) async {
+    await _lock.synchronized(() {
+      _cache.remove(key);
+    });
+  }
+
   /// Clears all data stored in the cache.
   ///
   /// - Removes all keys and values from the cache.